Updated plotCumulativeFractionOfReads
Changed plotCumulativeFractionOfReads to consider all cells for calculating cumsum because subsetting leads to different cumulative fractions based on the background considered and did not reflect real cumulative fraction or all aligned reads. Cutoff is still used for inflection point calculation and still leads to overestimation if high cutoffs are chosen!